Remove VERSIONS registry
When getting the schema registry, the passed in collection registry
needs to have all the vanilla registries added. So we have to separate
the two steps of getting collections and getting schemas.

I'm not sure if there's a better way other than exporting two functions
in java-1.16's index.ts file.